+++ app/storage/migrations/20260817182832_add_insurer_supports_instant_onboarding/migration.sql
@@ -1,0 +1,2 @@
+-- AlterTable
+ALTER TABLE "Insurer" ADD COLUMN     "supportsInstantOnboarding" BOOLEAN NOT NULL DEFAULT false;